==> train/1-build-dataset-spm.ipynb <==
import numpy as numpy
import pandas as pd
import matplotlib.pyplot as plt

normal_data = pd.read_json("normal.json")
chat_data = pd.read_json("chat.json")
group1_data = pd.read_json("clean-group-1.json")

spam_data = pd.read_json("spam.json")
spam_data2 = pd.read_json("spam2.json")
spam_data2_up = pd.read_json("spam2-up.json")

group1_df = pd.DataFrame({ 'type': 0, 'message': group1_data.iloc[:, 0] }) # (개인) 정상
normal_df = pd.DataFrame({ 'type': 0, 'message': normal_data.iloc[:, 0] }) # (경찰청) 광고
sampled_normal_df = normal_df.sample(n=5000, random_state=42)
chat_df = pd.DataFrame({ 'type': 0, 'message': chat_data.iloc[:, 0] }) # 한국어 멀티턴 대화
sampled_chat_df = chat_df.sample(n=5000, random_state=42)

spam_df = pd.DataFrame({ 'type': 1, 'message': spam_data.iloc[:4000, 0] }) # (경찰청) 스팸
spam2_df = pd.DataFrame({ 'type': 1, 'message': spam_data2.iloc[:, 0] }) # (개인) 스팸
spam2_up_df = pd.DataFrame({ 'type': 1, 'message': spam_data2_up.iloc[:, 0] }) # (개인) 스팸 증강

print(sampled_normal_df.shape[0])
print(sampled_chat_df.shape[0])
print(group1_df.shape[0])
print(spam_df.shape[0])
print(spam2_df.shape[0])

data = pd.concat([normal_df, chat_df, group1_df, spam_df, spam2_df, spam2_up_df], ignore_index=True)

with open("dataset.txt", "w", encoding="utf-8") as f:
    for message in data["message"]:
        f.write(str(message) + "\n")

from sklearn.model_selection import train_test_split

data = pd.concat([sampled_normal_df, sampled_chat_df, group1_df, spam_df, spam2_df, spam2_up_df], ignore_index=True)

X = data["message"]
y = data["type"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

train_df = pd.DataFrame({'message': X_train, 'type': y_train})
train_df.to_json('dataset-train-spm.json', orient='records', indent=4, force_ascii=False)

test_df = pd.DataFrame({'message': X_test, 'type': y_test})
test_df.to_json('dataset-test-spm.json', orient='records', indent=4, force_ascii=False)
