==> train/2-train-cnn-spm.ipynb <==
from tokenizers import Tokenizer

tokenizer = Tokenizer.from_pretrained("LGAI-EXAONE/EXAONE-3.5-7.8B-Instruct")

from pandas import pandas as pd

data_train = pd.read_json("dataset-train-spm.json")
data_train["ids"] = data_train["message"].apply(lambda x: tokenizer.encode(x).ids)

X_train = data_train["ids"]
Y_train = data_train["type"]
tokenizer.get_vocab_size()

data_train

from tf_keras.preprocessing.sequence import pad_sequences

max_len = 200
X_train = pad_sequences(X_train, maxlen=max_len)

from tf_keras.layers import Dense, Conv1D, GlobalMaxPooling1D, Embedding, Dropout, MaxPooling1D
from tf_keras.models import Sequential
from tf_keras.callbacks import EarlyStopping, ModelCheckpoint

embedding_dim = 32
num_filters = 128
kernel_size = 20
dropout_ratio = 0.3

model = Sequential()
model.add(Embedding(tokenizer.get_vocab_size(), embedding_dim))
model.add(Dropout(dropout_ratio))
model.add(Conv1D(num_filters, kernel_size, padding='valid', activation='relu'))
model.add(GlobalMaxPooling1D())
model.add(Dropout(dropout_ratio))
model.add(Dense(1, activation='sigmoid'))
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc'])

es = EarlyStopping(monitor='val_loss', mode='min', verbose=1, patience=3)
mc = ModelCheckpoint('cnn-spm.keras', monitor = 'val_acc', mode='max', verbose=1, save_best_only=True)

history = model.fit(X_train, Y_train, epochs=20, batch_size=64, validation_split=0.2, callbacks=[es, mc])
