==> train/2-train-freq.ipynb <==
from pandas import pandas as pd

df = pd.read_json("dataset-train.json")

spam = df[df["type"] == 1]

all_tokens = []
for tokens in spam["tokens"]:
    all_tokens.extend(tokens)

from collections import Counter

token_counts = Counter(all_tokens)
sorted_token_counts = token_counts.most_common()
sorted_token_counts
