==> train/3-query-rnn.ipynb <==
import pickle
from tf_keras.preprocessing.text import Tokenizer

with open('tokenizer.pickle', 'rb') as handle:
    tokenizer = pickle.load(handle)

from tf_keras.models import load_model
model = load_model('rnn.keras')

from kiwipiepy import Kiwi
from pandas import pandas as pd
from cleanutils import clean_message

kiwi = Kiwi()

def get_tokens(text):
    if pd.isna(text) or not text:
        return []
    text = clean_message(text, False)
    tokens = kiwi.tokenize(str(text))
    filtered_tokens = []
    for token in tokens:
        if token.tag[0] in ['N', 'V', 'M', 'S', 'U']:
            filtered_tokens.append(token.form)
    return filtered_tokens

queries = [
    "[대한통운] 19523523912 금일 13시 배송예정. 담당기사 박* (*-*-*))",
    "무엇이 기다리고 있나요? 저는 아파트에 배송 했어요.",
    "신!규3만-지-급 출1금OK 첫 츙+ 뽀5너스 1000+",
    "안녕하세요, 고객님. 신규 회원가입을 축하드립니다. 추가 3만원 쿠폰을 받으려면?",
    "축하합니다 고객님! 문자를 받으신 분께만 깜짝 10,000원 적립금을 드렸어요.",
    "tel: vji32d 안녕하세요 고객님, 문자를 받으신 분들은 적립금을 드리니 확인해 보세요."
]

tokens = [get_tokens(q) for q in queries]
print(tokens)
tokens = tokenizer.texts_to_sequences(tokens)

max_len = 200
from tf_keras.preprocessing.sequence import pad_sequences
tokens = pad_sequences(tokens, maxlen=max_len)

results = model.predict(tokens)
for i, probability in enumerate(results):
    percentage = probability[0] * 100
    print(f"{percentage:.0f}%, {queries[i]}")
